# Giorno 13/Esercizio 2/api-docs.customization.api-spec.json
{"openapi":"3.1.0","info":{"title":"API Studenti","description":"API REST per la gestione degli studenti","contact":{"name":"Francesco Cerrato","email":"Francescocerrato200@email.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Studenti","description":"Gestione studenti"}],"paths":{"/api/studenti":{"get":{"tags":["Studenti"],"summary":"Restituisce l'elenco degli studenti e informazioni Pageable","operationId":"getStudenti","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Studente trovato","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageStudente"}}}},"404":{"description":"Studente non trovato","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageStudente"}}}}}},"post":{"tags":["Studenti"],"summary":"metodo POST per creare/salvare nuovi studenti nel db","operationId":"creaStudente","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Studente"}}},"required":true},"responses":{"201":{"description":"Studente creato","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Studente"}}}},"400":{"description":"Richiesta non valida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Studente"}}}}}}},"/api/studenti/{id}":{"get":{"tags":["Studenti"],"summary":"Restituisce uno studente dato il suo id","operationId":"StudenteByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Studente trovato","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Studente"}}}},"404":{"description":"Studente non trovato","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Studente"}}}}}},"delete":{"tags":["Studenti"],"summary":"Metodo DELETE per eliminare uno studente, dato il suo id, dal DB","operationId":"eliminaStudente","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Eliminato con successo"},"404":{"description":"Studente non trovato"}}},"patch":{"tags":["Studenti"],"summary":"Esegue la modifica parziale di uno Studente","operationId":"patchStudente","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Studente"}}}}}}},"/api/studenti/{id}/corso":{"get":{"tags":["Studenti"],"summary":"Restituisce il corso di laurea di uno specifico studente","operationId":"CercaCorsoLaureaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/studenti/corso/{corso}":{"get":{"tags":["Studenti"],"summary":"Restituisce ogni Studente di uno specifico corso","operationId":"cercaPerCorso","parameters":[{"name":"corso","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Studente trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}},"404":{"description":"Studente non trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}}}}},"/api/studenti/cognome/{cognome}":{"get":{"tags":["Studenti"],"summary":"Restituisce ogni Studente con uno specifico cognome","operationId":"cercaPerCognome","parameters":[{"name":"cognome","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Studente trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}},"404":{"description":"Studente non trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}}}}},"/api/studenti/cercaxnomecontaining":{"get":{"tags":["Studenti"],"summary":"Restituisce ogni Studente il cui nome contiene una specifica stringa","operationId":"cercaPerNomeContaining","parameters":[{"name":"parte","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Studente trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}},"404":{"description":"Studente non trovato","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}}}}}}}}},"components":{"schemas":{"Studente":{"type":"object","description":"Tabella studenti","properties":{"id":{"type":"integer","format":"int32","description":"Id univoco dello studente","example":1},"nome":{"type":"string","description":"Nome dello studente","example":"Mario"},"cognome":{"type":"string","description":"Cognome dello studente","example":"Aliberti"},"email":{"type":"string","description":"Email dello studente","example":"Mario.Rossi@gmail.com"},"data_nascita":{"type":"string","format":"date","description":"Data di nascita dello studente","example":"1998-10-10"},"getcorsolaurea":{"type":"string","description":"Corso di laurea dello studente","example":"Architettura"},"corso_laurea":{"type":"string","writeOnly":true}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageStudente":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Studente"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}}}}}